Create README.md
This was a task for a project during my internship with Genesys Tech Hub, Do not mind the UI , it wasn't graded because this was a Backend Task and what mattered is the Javascript code. Decided to upload it because it was really tasking and i learnt a whole lot while at it.